# Conflicts:
#	gms-server/src/main/resources/application.yml

+++ gms-server/src/main/java/org/gms/dao/entity/PlayernpcsEquipDO.java
@@ -36,6 +36,6 @@ public class PlayernpcsEquipDO implements Serializable {
 
     private Integer type;
 
-    private Integer equippos;
+    private Short equippos;
 
 }
